Chnaged imports from 'useful_functions' to 'utils' and move clean_grants_data from grant data processing to utils and correct imports accordingly
